requirements.txt
groq>=0.4.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
streamlit>=1.30.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
torch>=2.0.0
